pyproject.toml
[project]
name = "amdis-env"
version = "0.1.0"
description = "AMDIS RL Environment"
authors = [
    {name = "User"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "openenv-core (>=0.2.3,<0.3.0)"
]

[project.scripts]
server = "microbiome.server.app:main"

[tool.poetry]
packages = [
    { include = "microbiome" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
